Create selectionSort.c
Selection Sort.

Selection Sort is a very bad algorithm to be honest but it's very often asked in exams so you should know it. 
Things work simply in selection sort. We select a item say at index zero and we loop through the array from the next element. If any
element is found to be greater than the first element, then we perform a swap. Repeat this for all current elements. 


Algorithm For selection sort
Step 1: Take a array, initialize two loop counters i and j.
Step 2: For every element at index i, the program loops through the array from the i+1 index (we use j for this)
        If a greater number is found, then we perform a basic swap using a temp variable. 
        Repeat this procedure for the remaining items.